Fix various 404s in the RSS feeds
The Jekyll variable site.url is only used in a few places and, on the live site, is set to http://seananderson.ca/blog.html. I believe this should be set to http://seananderson.ca instead.

The places where the variable site.url is used end up with links that 404. I've done my best to glance at the code to make sure this doesn't break anything else. Might be worth a once-over if you plan to merge.